diti_server/requirements.txt
APScheduler==4.0.0a5
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.3
defusedxml==0.8.0rc2
diff-match-patch==20241021
Django==5.1.2
django-allauth==65.1.0
django-cors-headers==4.5.0
django_crontab==0.7.1
django-dbbackup==4.2.1
django-env-overrides==0.1.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.2.0
django-jazzmin==3.0.1
django-jsonfield==1.4.1
django-mass-edit==3.5.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==2.0.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
MarkupPy==1.14
MarkupSafe==3.0.2
numpy==2.1.2
oauthlib==3.2.2
odfpy==1.4.1
openapi-codec==1.3.2
opencv-python==4.10.0.84
openpyxl==3.2.0b1
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyopenssl==24.2.1
pyotp==2.9.0
Python-IO==0.3
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
scipy==1.14.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.1
tablib==3.7.0
tzlocal==5.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
werkzeug==3.0.6
xlrd==2.0.1
xlwt==1.3.0
